requirements.txt
pip==21.1.3
Django==2.2.24
mysqlclient==2.0.3

wagtail==2.13.4
pytest-django
django-compressor==2.4.1
django-extensions==3.1.3
ipython==7.25.0
whitenoise==5.3.0
diskcache==5.2.1
redis==3.5.3
django-rq==2.4.1
rq-scheduler==0.11.0
django-webpack-loader==1.1.0
geoip2==4.2.0

django-filter
django-crispy-forms
django-debug-toolbar
selenium
pandas
lxml
django-markdownx
sorl-thumbnail
dj-pagination
django-autocomplete-light
python-geoip
python-geoip-geolite2

git+https://github.com/fruitschen/django-feedreader.git#egg=django-feedreader
-e git+https://github.com/django-backup/django-backup.git@develop#egg=django-backup